crida/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-bcrypt==1.0.1
PyJWT==2.8.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
fpdf2==2.7.9
Pillow==10.3.0
rembg==2.0.57
face-recognition==1.3.0
numpy==1.26.4
